refactor: move coercion functions from schema.py to udf
fixes #3033

BREAKING CHANGE: coercion functions previously in expr/schema.py are now in udf/vectorized.py
